adds stub for dm api subcontractor endpoint
