+++ conf/resources/data/dm/subcontractor/wmfId.json
@@ -1,0 +1,110 @@
+[
+  {
+    "fileInd": "K",
+    "wmfId": "00000000000036320200100009KP0001",
+    "shortDescription": "RGD Retrn Chrg",
+    "solDescription": "Interest on First Payment on Account",
+    "latestDueDate": "2021-09-15",
+    "accountingPeriodEndDate": "2021-09-15",
+    "safeChargeRef": "XA188698553712",
+    "customerRefType": "A",
+    "outputRefType": "B",
+    "AdditionalInfo": [
+      {
+        "infoType": "RCQ",
+        "infoValue": "Reinstated Charge"
+      }
+    ],
+    "bsPeriod": [
+      {
+        "irIntAreaRef": "AX21231YE"
+      }
+    ],
+    "debtItemSignal": [
+      {
+        "signalType": "#TBD#",
+        "signalValue": "C"
+      },
+      {
+        "signalType": "#TBD#",
+        "signalValue": "D"
+      }
+    ],
+    "debtItemCharge": [
+      {
+        "intBearingInd": "Y",
+        "interestChargeInd": "Y",
+        "currentCollectibleAmount": 34409.49,
+        "chargeAmount": 7899.45,
+        "interestStartDate": "2021-09-15",
+        "interestSuppressionInd": "Y",
+        "debtChargePayment": [
+          {
+            "postingAmount": 100.99,
+            "effectiveDateOfPayment": "2021-09-15",
+            "adjustmentType": "CHG"
+          },
+          {
+            "postingAmount": 110.99,
+            "effectiveDateOfPayment": "2021-10-15",
+            "adjustmentType": "CHG"
+          }
+        ]
+      }
+    ]
+  },
+  {
+    "fileInd": "L",
+    "wmfId": "00000000000036320200100009KP0002",
+    "shortDescription": "RGD Retrn Chrg",
+    "solDescription": "Interest on First Payment on Account",
+    "latestDueDate": "2021-10-15",
+    "accountingPeriodEndDate": "2021-10-15",
+    "safeChargeRef": "XA188698553712",
+    "customerRefType": "A",
+    "outputRefType": "B",
+    "AdditionalInfo": [
+      {
+        "infoType": "RCQ",
+        "infoValue": "Reinstated Charge"
+      }
+    ],
+    "bsPeriod": [
+      {
+        "irIntAreaRef": "AX21231YE"
+      }
+    ],
+    "debtItemSignal": [
+      {
+        "signalType": "#TBD#",
+        "signalValue": "D"
+      },
+      {
+        "signalType": "#TBD#",
+        "signalValue": "E"
+      }
+    ],
+    "debtItemCharge": [
+      {
+        "intBearingInd": "Y",
+        "interestChargeInd": "Y",
+        "currentCollectibleAmount": 54409.49,
+        "chargeAmount": 8899.45,
+        "interestStartDate": "2021-10-15",
+        "interestSuppressionInd": "Y",
+        "debtChargePayment": [
+          {
+            "postingAmount": 200.99,
+            "effectiveDateOfPayment": "2021-10-15",
+            "adjustmentType": "CHG"
+          },
+          {
+            "postingAmount": 210.99,
+            "effectiveDateOfPayment": "2021-11-15",
+            "adjustmentType": "CHG"
+          }
+        ]
+      }
+    ]
+  }
+]
